Auto detect tables for fast snapshot restore
Now all the tables inside a database will be included as part of CDC.
This is required becuase of fast snapshot restore mechanism, since we do
byte by byte copy instead of running hour long transactions in batches
to restore a snapshot of database. An endpoint has also been exposed to
post a database as multi-part request. This is experimental right now
and might be used in future to have boot sequence to download and start
database file from remote resource.