Adding support for snapshotting
This will allow Marmot to save and restore snapshot for the database.
Right now the snapshot shard is 1, meaning that a snapshot is taken if
log entries of shard 1 hit MaxLogEntries `max-log-entries`. This is not
a very good deterministic way of taking snapshot we might need a better
way to decide when a snapshot is taken. But for now assuming the hashing
function will distribute logs uniformly, on average these logs will be
pretty close, and early snapshot won't hurt.