Fixing issue for multi-PK tables
Turns out SQLite schema returns index of the PK in table rather than a
boolean (a very rational choice IMO). Looking at code turns out I was
dumb enough to not read through the docs and map it to a boolean
instead. Now this can be a bug in ORM mapper mapping anything other than
1 to false but the right way to do is to actually read it out in an
  integer and then map it to boolean.

I am not very happy with current fix but in order to keep things simple
this should be good enough fix.